Refinements and cleanup following in-person code review
